Фильтр отрисовывается вместе с меню
